chat_server/chat/requirements.txt
llama-index == 0.10.20
google-auth == 2.29.0
Flask == 3.0.0
openai == 1.14.2
selenium == 4.19.0
pdfkit == 1.0.0
pywebcopy == 7.0.2
beautifulsoup4 == 4.12.3
PyPDF2 == 3.0.1
google-auth-oauthlib == 1.2.0
google-api-python-client == 2.127.0
llama-index-embeddings-huggingface == 0.1.4
llama-index-llms-ollama == 0.1.2
llama-index-readers-web == 0.1.8
webdriver-manager == 4.0.1
Flask-Cors == 4.0.0
sentence-transformers == 2.7.0
